Move process name extraction to pkg. Add PID to data returned from process files extraction.
