+++ lib/app/nagiosfoundation/check_process_linux.go
@@ -2,10 +2,12 @@
 
 package nagiosfoundation
 
+import "github.com/ncr-devops-platform/nagiosfoundation/lib/pkg/process"
+
 func isProcessRunningOsConstrained(name string) bool {
 	retVal := false
 
-	if processEntries, _ := getProcessesByName(name); len(processEntries) > 0 {
+	if processEntries, _ := process.GetProcessesByName(name); len(processEntries) > 0 {
 		retVal = true
 	}
